requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
Django==3.2.9
django-rest-auth==0.9.5
djangorestframework==3.12.4
future==0.18.2
idna==3.3
mando==0.6.4
Pillow==8.4.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7